+++ src/data/databaseMetadata.ts
@@ -1,7 +1,7 @@
 export const databaseMetadata = {
-  databaseVersion: '0.23.2',
+  databaseVersion: '0.23.3',
   schemaVersion: 13,
-  lastUpdated: '2026-08-01',
+  lastUpdated: '2026-08-02',
   officialRosterLastChecked: '2026-06-23',
   gameVersion: '26.6.53509',
   currentDocumentedGameBuild: '26.6.53509',